dataset/requirements.txt
certifi==2020.12.5
lxml==4.6.3
numpy==1.20.2
Pillow==8.2.0
progressbar2==3.53.1
python-utils==2.5.6
six==1.15.0
wincertstore==0.2
